Refactored folder structure and component naming
