+++ src/Pages/NotFound/NotFound.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Navbar from '../Components/Navbar';
+import Navbar from '../../Components/Navbar/Navbar';
 import { Link } from 'react-router-dom';
 
 function NotFound() {
